requirements.txt
Flask
Werkzeug
openai
python-docx
docx2txt
flask-socketio
Werkzeug
tenacity
